requirements.txt
requests
binance-connector
time